Create a common view module and inherit Statuses and Countries views from it
